refactor(all): 우테코 코드스타일 적용. 패키지명 변경 model -> domain
