+++ src/main/java/racingcar/domain/RandomUtil.java
@@ -1,11 +1,12 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.Random;
 
 public class RandomUtil {
     private static final Random random = new Random();
 
-    private RandomUtil() {}
+    private RandomUtil() {
+    }
 
     public static int getNumbersInRange(final int endInclusive) {
         return random.nextInt(endInclusive);
