+++ src/main/java/racingcar/domain/vo/MovableStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.domain.vo;
+
+public interface MovableStrategy {
+    boolean isMove();
+}
