--- src/main/java/racingcar/model/vo/MovableStrategy.java
@@ -1,5 +1,0 @@
-package racingcar.model.vo;
-
-public interface MovableStrategy {
-	boolean isMove();
-}
