+++ src/main/java/racingcar/view/OutputView.java
@@ -1,15 +1,15 @@
 package racingcar.view;
 
 import java.util.List;
-
-import racingcar.model.vo.Car;
+import racingcar.domain.vo.Car;
 
 public class OutputView {
     private static final String WINNERS_DELIMITER = ", ";
     private static final String WINNER_ANNOUNCE_MESSAGE = "%s가 최종 우승했습니다.";
     private static final String RACE_PROGRESS_CHARACTER = "-";
 
-    private OutputView() {}
+    private OutputView() {
+    }
 
     public static void printCurrentPosition(final List<Car> cars) {
         cars.forEach(OutputView::printRacingProgress);
