+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,12 +1,14 @@
 package calculator;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+
 import calcaulator.StringCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import static org.assertj.core.api.Assertions.*;
 
 public class StringCalculatorTest {
     @ParameterizedTest
@@ -49,6 +51,6 @@ public class StringCalculatorTest {
     @DisplayName("음수를 전달할 경우 RuntimeException 예외가 발생해야 한다")
     void throwRuntimeExceptionWhenMinus(String text) {
         assertThatExceptionOfType(RuntimeException.class)
-            .isThrownBy(() -> StringCalculator.addString(text));
+                .isThrownBy(() -> StringCalculator.addString(text));
     }
 }
