+++ src/test/java/racingcar/domain/RandomUtilTest.java
@@ -1,9 +1,8 @@
-package racingcar.model;
+package racingcar.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.IntStream;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,9 +11,9 @@ public class RandomUtilTest {
     @DisplayName("0과 9사이의 숫자 랜덤 생성 테스트")
     void generateRandomNumber() {
         IntStream.range(0, 10000)
-            .forEach(i -> {
-                int result = RandomUtil.getNumbersInRange(10);
-                assertThat(result < 10 && result >= 0).isTrue();
-            });
+                .forEach(i -> {
+                    int result = RandomUtil.getNumbersInRange(10);
+                    assertThat(result < 10 && result >= 0).isTrue();
+                });
     }
 }
